fix error in validatoin

+++ app/Services/PortfoliosService.php
@@ -83,19 +83,24 @@ class PortfoliosService
             $path = $data['preview']->storeAs('portfolio/' . $portfolio->id, $imageName, 'public');
             $portfolio->update(['preview' => $path]);
         }
-        if (isset($data['images'])) {
-            $files = Storage::files('public/' . $portfolio->images);
-            foreach ($files as $file) {
-                if ($file !== 'public/' . $portfolio->preview) {
-                    Storage::delete($file);
-                }
+
+
+        $existImages = $data['existImages'] ?? [];
+        $files = Storage::files('public/' . $portfolio->image);
+        foreach ($files as $file) {
+            if ($file !== 'public/' . $portfolio->preview && !in_array($file, $existImages)) {
+                Storage::delete($file);
             }
-            foreach ($data['images'] as $image) {
-                $imageName = $image->getClientOriginalName();
-                $image->storeAs('portfolio/' . $portfolio->id, $imageName, 'public');
-            }
-            $portfolio->update(['images' => 'portfolio/' . $portfolio->id]);
         }
+
+        if (isset($data['images'])) {
+            foreach ($data['image'] as $image) {
+                $imageName = $image->getClientOriginalName();
+                $image->storeAs('service/' . $portfolio->id, $imageName, 'public');
+            }
+            $portfolio->update(['images' => 'service/' . $portfolio->id]);
+        }
+
         unset($data['preview']);
         unset($data['images']);
         $portfolio->update($data);
